extract exit practice to standalone button
